feat: add /save endpoint for events

+++ eventify/src/main/java/com/eventify/user/repository/UserRepository.java
@@ -8,6 +8,6 @@ import org.springframework.stereotype.Repository;
 import com.eventify.user.model.User;
 
 @Repository
-public interface UserRepository extends JpaRepository<User, Integer> {
+public interface UserRepository extends JpaRepository<User, Long> {
     Optional<User> findByEmail(String email);
 }
